Rename remaining safe assistant filenames
