+++ apps/agor-docs/content/blog/agor-assistants.mdx
@@ -2,7 +2,7 @@
 title: 'Introducing Agor Teammates: Persistent AI Companions for Developer Teams'
 description: 'What started as an OpenClaw experiment is now a first-class Agor feature. Meet Teammates — persistent AI entities with memory, skills, and team-wide reach through Slack.'
 date: 2026-03-03
-image: '/images/blog/agor-assistants.png'
+image: '/images/blog/agor-teammates.png'
 ---
 
 # Introducing Agor Teammates
