+++ apps/agor-docs/content/blog/claude-tag-vs-agor-assistants.mdx
@@ -2,7 +2,7 @@
 title: 'Claude Tag Validates the Category. Now Model Your Teammates.'
 description: 'Claude Tag brings @Claude into Slack. Here is why teams will still need modeled, governed, observable, multi-specialist teammates — and why Agor is built for that next step.'
 date: 2026-06-26
-image: '/images/blog/claude-tag-vs-agor-assistants.png'
+image: '/images/blog/claude-tag-vs-agor-teammates.png'
 ---
 
 # Claude Tag Validates the Category. Now Model Your Teammates.
