+++ apps/agor-docs/public/images/blog/agor-teammates.png
