+++ packages/core/src/templates/teammate-welcome-note-compat.ts
