+++ packages/core/tsup.config.ts
@@ -24,7 +24,7 @@ export default defineConfig({
     'templates/session-context': 'src/templates/session-context.ts', // Agor system prompt rendering
     'templates/spawn-subsession-template': 'src/templates/spawn-subsession-template.ts', // Spawn-subsession meta-prompt
     'templates/teammate-welcome-note': 'src/templates/teammate-welcome-note.ts', // Teammate board welcome note renderer
-    'templates/assistant-welcome-note': 'src/templates/assistant-welcome-note.ts', // Deprecated assistant board welcome note compatibility renderer
+    'templates/assistant-welcome-note': 'src/templates/teammate-welcome-note-compat.ts', // Deprecated compatibility renderer for the old public export path
     'templates/zone-trigger-context': 'src/templates/zone-trigger-context.ts', // Canonical zone-trigger context builder
     'environment/variable-resolver': 'src/environment/variable-resolver.ts', // Environment variable resolution
     'environment/render-snapshot': 'src/environment/render-snapshot.ts', // v2 branch env snapshot rendering
